geo/mesh:
Added handling for a very obscure case which we can encounter trying to
`_get_proj_enclosing_polygon` on a RectangularMesh.

The comments from this patch read:

NOTE(larsbutler): In some rare cases, we've observed ValueErrors
("No Shapely geometry can be created from null value") with very
specific sets of polygons such that there are two unique
and many duplicates of one.
This bug is very difficult to reproduce consistently (except on
specific platforms) so the work around here is to remove the
duplicate polygons. In fact, we only observed this error on our
CI/build machine. None of our dev environments or production
machines has encountered this error, at least consistently. >:(